--- demo/annotator_test/index.html
@@ -1,64 +1,0 @@
-<html>
-<head>
-<link rel=”stylesheet” href='annotator.1.2.10/annotator.min.css'>
-</head>
-<body>
-<script src='http://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js'></script>
-<script src='annotator.1.2.10/annotator.min.js'></script>
-<script type='text/javascript'>
-jQuery(function ($) {
-    var content = $('#content').annotator();
-    content.annotator('setupPlugins');
-
-    // Annotator.Plugin.StoreLogger = function (element) {
-    //     return {
-    //         pluginInit: function () {
-    //         this.annotator
-    //             .subscribe("annotationCreated", function (annotation) {
-    //                 console.info("The annotation: %o has just been created!", annotation)
-    //             })
-    //             .subscribe("annotationUpdated", function (annotation) {
-    //                 console.info("The annotation: %o has just been updated!", annotation)
-    //             })
-    //             .subscribe("annotationDeleted", function (annotation) {
-    //                 console.info("The annotation: %o has just been deleted!", annotation)
-    //             });
-    //         }
-    //     }
-    // };
-
-    // content.annotator('addPlugin', 'StoreLogger');
-
-});
-</script>
-<div id='content'>
-<h1>HTML Ipsum Presents</h1>
-
-<p><strong>Pellentesque habitant morbi tristique</strong> senectus et netus et malesuada fames ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam egestas semper. <em>Aenean ultricies mi vitae est.</em> Mauris placerat eleifend leo. Quisque sit amet est et sapien ullamcorper pharetra. Vestibulum erat wisi, condimentum sed, <code>commodo vitae</code>, ornare sit amet, wisi. Aenean fermentum, elit eget tincidunt condimentum, eros ipsum rutrum orci, sagittis tempus lacus enim ac dui. <a href="#">Donec non enim</a> in turpis pulvinar facilisis. Ut felis.</p>
-
-<h2>Header Level 2</h2>
-
-<ol>
-   <li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li>
-   <li>Aliquam tincidunt mauris eu risus.</li>
-</ol>
-
-<blockquote><p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus magna. Cras in mi at felis aliquet congue. Ut a est eget ligula molestie gravida. Curabitur massa. Donec eleifend, libero at sagittis mollis, tellus est malesuada tellus, at luctus turpis elit sit amet quam. Vivamus pretium ornare est.</p></blockquote>
-
-<h3>Header Level 3</h3>
-
-<ul>
-   <li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li>
-   <li>Aliquam tincidunt mauris eu risus.</li>
-</ul>
-
-<pre><code>
-#header h1 a {
-  display: block;
-  width: 300px;
-  height: 80px;
-}
-</code></pre>
-</div>
-</body>
-</html>
