--- demo/annotator_test/test.js
@@ -1,1 +1,0 @@
-alert("hello world")
