--- demo/mockScholarship/MockScholarship.html
@@ -1,56 +1,0 @@
-<html>
-    <head>
-        <title>Poly Publishing</title>
-        <link rel="stylesheet" href="style.css">
-    </head>
-    <body>
-        <div class="content">
-            <div class="sidebar">
-                <h2>Poly Publishing</h2>
-                <div>
-                    <h4>Digital Publications</h4>
-                    <ul>
-                        <li>Objects of Affection</li>
-                        <li>Living Library</li>
-                        <li>#IAmCalPoly</li>
-                    </ul>
-                </div>
-                <div>
-                    <h4>Create Your Own</h4>
-                    <ul>
-                        <li>Faculty Exhibit Program</li>
-                        <li>Reviewing & Contributing</li>
-                        <li>Digital Tools</li>
-                        <li>Consultations</li>
-                    </ul>
-                </div>
-                <div>
-                    <h4>About Us</h4>
-                    <ul>
-                        <li>Creative Works</li>
-                        <li>Sustainability Plans</li>
-                        <li>Partner with Us</li>
-                        <li>References & Credits</li>
-                    </ul>
-                </div>
-                <div>
-                    <h4>Last Updated</h4>
-                </div>
-                <div>
-                    <p>
-                        Though this is a platform for scholarship that hopes to promote open access to knowledge, the Creative Works team fully acknowledges that simly stating 'open access' does not change the systems of institutionally, inequality, and colonialism that higher education standards are built upon. This resource hopes to lessen the digital divide by access to knowledge, but is not directly changing the material inequalities that the digital divide is built upon. 
-                    </p>
-                </div>
-            </div>
-            <div class="scholarship">
-                <ul class="navbar">
-                    <li>Experience</li>
-                    <li class="active">About</li>
-                    <li>Contribute</li>
-                    <li>Revisions</li>
-                </ul>
-                <img src="objectsOfAffection.png">
-            </div>
-        </div>
-    </body>
-</html>
