+++ package.json
@@ -12,7 +12,7 @@
     "@types/react-dom": "^16.9.8",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.3",
+    "react-scripts": "^4.0.0",
     "typescript": "^4.0.3"
   },
   "scripts": {
