+++ public/index.html
@@ -1,44 +1,67 @@
-﻿<html>
-<head>
-    <meta charset="utf-8">
-    <title></title>
-</head>
-<body>
-    <h1>HTML Ipsum Presents</h1>
-    <div class="well">
-        <h1>Sample app</h1>
-        <div id="root"></div>
-    </div>
-    
-    <p>
-        <strong>Pellentesque habitant morbi tristique</strong> 
-        senectus et netus et malesuada fames ac turpis egestas. Vestibulum tortor quam, feugiat vitae, ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam egestas semper. 
-        <em>Aenean ultricies mi vitae est.</em> 
-        Mauris placerat eleifend leo. Quisque sit amet est et sapien ullamcorper pharetra. Vestibulum erat wisi, condimentum sed, 
-        <code>commodo vitae</code>
-        , ornare sit amet, wisi. Aenean fermentum, elit eget tincidunt condimentum, eros ipsum rutrum orci, sagittis tempus lacus enim ac dui. 
-        <a href="#">Donec non enim</a> 
-        in turpis pulvinar  facilisis. Ut felis.
-    </p>
-    
-    <h2>Header Level 2</h2>
-    
-    <ol>
-       <li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li>
-       <li>Aliquam tincidunt mauris eu risus.</li>
-    </ol>
-    
-    <blockquote>
-        <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus magna. Cras in mi at felis aliquet congue. Ut a est eget ligula molestie gravida. Curabitur massa. Donec eleifend, libero at sagittis mollis, tellus est malesuada tellus, at luctus turpis elit sit amet quam. Vivamus pretium ornare est.
-        </p>
-    </blockquote>
-    
-    <h3>Header Level 3</h3>
-    
-    <ul>
-        <li>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</li>
-        <li>Aliquam tincidunt mauris eu risus.</li>
-    </ul>
-</body>
+<html>
+    <head>
+        <title>Poly Publishing</title>
+        <link rel="stylesheet" href="style.css">
+    </head>
+    <body>
+        </div>
+        <div class="content">
+            <div id="root"></div>
+            <div class="sidebar">
+                <h2>Poly Publishing</h2>
+                <div>
+                    <h4>Digital Publications</h4>
+                    <ul>
+                        <li>Objects of Affection</li>
+                        <li>Living Library</li>
+                        <li>#IAmCalPoly</li>
+                    </ul>
+                </div>
+                <div>
+                    <h4>Create Your Own</h4>
+                    <ul>
+                        <li>Faculty Exhibit Program</li>
+                        <li>Reviewing & Contributing</li>
+                        <li>Digital Tools</li>
+                        <li>Consultations</li>
+                    </ul>
+                </div>
+                <div>
+                    <h4>About Us</h4>
+                    <ul>
+                        <li>Creative Works</li>
+                        <li>Sustainability Plans</li>
+                        <li>Partner with Us</li>
+                        <li>References & Credits</li>
+                    </ul>
+                </div>
+                <div>
+                    <h4>Last Updated</h4>
+                </div>
+                <div>
+                    <p>
+                        Though this is a platform for scholarship that hopes to promote open access to knowledge, the Creative Works team fully acknowledges that simly stating 'open access' does not change the systems of institutionally, inequality, and colonialism that higher education standards are built upon. This resource hopes to lessen the digital divide by access to knowledge, but is not directly changing the material inequalities that the digital divide is built upon. 
+                    </p>
+                </div>
+            </div>
+            <div class="scholarship">
+                <ul class="navbar">
+                    <li>Experience</li>
+                    <li class="active">About</li>
+                    <li>Contribute</li>
+                    <li>Revisions</li>
+                </ul>
+                <img src="objectsOfAffection.png">
+                <div class="scholarship text1">
+                    <h4> Just a random paragraph to try to test the highlighting feature with children</h4>
+                </div>
+                <div class="scholarship text2">
+                    This is a non-bolded portion of a paragraph like when users write paragraphs about their work. They could go on for quite a bit if they wanted to.
+                </div>
+                <div class="scholarship text3">
+                    And here is the final line where they would tie everything up with a nice neat bow.                
+                </div>
+            </div>
+        </div>
+    </body>
 </html>
