--- src/hello.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-export const HelloComponent = () => {
-  return (
-    <h2>Hello component !</h2>
-  );
-}
