+++ src/index.tsx
@@ -1,12 +1,19 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { HelloComponent } from './hello';
+import { HighlightButton } from './highlight';
 import { setup } from './setup';
+import { startHighlight } from './highlight';
 
 setup();
 ReactDOM.render(
-    <HelloComponent/>,
-    document.getElementById('root')
+    <HighlightButton/>,
+    document.getElementById('root')   
 );
+ checkHighlight();
+
+export function checkHighlight () {
+    let highlightBtn = document.getElementById("highlightBtn");
+    highlightBtn?.addEventListener("click", (e:Event) => startHighlight());
+}
